Feature/integrating all features (#34)
* Many formatting updatesgit add . woohoo

* Reorganized

* WIP: fixing linter

* linter fixed! lets goo

* we're so back (compile errors fixed)

* Fixed ethernet implementation

* fix: referenced mac address

* updated pedals can interface to include a heartbeat

* updated timeout to 20ms

* starting integration of pedals timeout and controls

* vcr recvs pedals data now

* removing wanting ready to drive state and adding in pedals timeout check to state machine

* adding in controller integration with tc mux min viable and mode 0 and the state machine

* minor fix to remove extranneous timestamp from stamped pedals data

* Fixed some linter errors

* asdf

* ferst drive

---------

+++ lib/systems/src/controllers/DrivebrainController.cpp
@@ -35,10 +35,10 @@ DrivetrainCommand_s DrivebrainController::evaluate(const VCRData_s &state, unsig
 
     if( (curr_millis - last_speed_setpoint_timestamp) > _worst_message_latencies.worst_speed_setpoint_latency_so_far)
     {
-        _worst_message_latencies.worst_speed_setpoint_latency_so_far = (curr_millis - last_speed_setpoint_timestamp);   
+        _worst_message_latencies.worst_speed_setpoint_latency_so_far = static_cast<int64_t>(curr_millis - last_speed_setpoint_timestamp);   
     } else if((curr_millis - last_torque_lim_timestamp) > _worst_message_latencies.worst_torque_lim_latency_so_far)
     {
-        _worst_message_latencies.worst_torque_lim_latency_so_far = (curr_millis - last_torque_lim_timestamp);   
+        _worst_message_latencies.worst_torque_lim_latency_so_far = static_cast<int64_t>(curr_millis - last_torque_lim_timestamp);   
     }
 
     bool timing_failure = (speed_setpoint_msg_too_latent || torque_limit_message_too_latent || not_all_messages_recvd || latency_diff_too_high);
